% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_cpdb_pathway_to_mgsa_geneset.R
\name{convert_cpdb_pathway_to_mgsa_geneset}
\alias{convert_cpdb_pathway_to_mgsa_geneset}
\title{Convert the CPDB Pathway Tab Format into MGSA Format}
\usage{
convert_cpdb_pathway_to_mgsa_geneset(infile, pathway.id.col = "external_id",
  gene.id.col = "hgnc_symbol_ids")
}
\arguments{
\item{infile}{Pathway to the CPDB pathway file}

\item{pathway.id.col}{Column name of the pathway identifer in the infile}

\item{gene.id.col}{Column name of the gene identifier in the infile}
}
\description{
This function converts the ConsensusPathDB \url{http://consensuspathdb.org/}
biological pathways (.tab) into a mgsa geneset object
}

